Update rain-world.md (#181)
* Update rain-world.md

Platinum status has been demoted to silver to reflect significant performance issues which require alternative solutions not found in-game. 
Adds a warning about performance dropping significantly depending on what creatures may be present and a solution in the form of a community made patch. It also addresses a common issue where upon launching the game and entering the Remix menu the game freezes, and how in most cases it can be fixed by entering the game and then leaving before opening the Remix menu. An additional note is left mainly as a recommendation or troubleshooting step and is not applicable in all cases but may be useful.

* Improve readibility of Rain World's Warning section

---------

Co-authored-by: mattmaniak <72657016+mattmaniak@users.noreply.github.com>